dev-requirements.txt
pytest-ckan
pytest-mock
pytest-factoryboy
black
pre-commit
